Add more tests fro response, with invalid requests. Rename some tests.

+++ src/common.rs
@@ -66,7 +66,7 @@ pub const MAX_HEADERS_AMOUNT: usize = 10_000;
 pub const MAX_REQUEST_BODY_SIZE: u64 = u64::MAX / 2; // 2 GB
 pub const MAX_HEADER_SIZE: u64 = (u16::MAX / 2) as u64; // 8 KB
 pub const DEFAULT_HTTP_VERSION: &str = "1.1";
-pub const MAX_URI_LENGTH: u16 = u16::MAX;
+pub const MAX_URI_LENGTH: usize = u16::MAX as usize;
 
 pub trait HttpStream: Read + Write {}
 impl<T: Read + Write> HttpStream for T {}
